The beginning of V0.2.Add new minimal sidebar navigation template. Add new minimal blog look. Upload for testing on mobile.
This is a new concept for the bones of v0.2 of my alessandrobisiani.github.io
It is ostensibly the starting point for the next version. It consists of one
page: the new index page. It is reachable through the "Sidebar Template Test
Page" link in the current nav bar. It doesn't use bootstrap. It includes a left
side navigation sidebar and new skinny blog look. Resizing to column layout
for "xs" screen sizes.
Minimal styling and no content.

Includes a short list of interesting colors, but the colors in use are
temporary.

Please enter the commit message for your changes. Lines starting